// FinalEmlak/EmlakUI/src/app/Models/Uyeler.ts
export class Uyeler{
        uid: string;                                               
        adsoyad: string;                                              
        sifre: string;                                              
        eposta: string;                                              
        uyeadmin: number;                                               
        kadi: string;
        foto:string;                                                                                            
}